requirements.txt
Django==5.0
Pillow==10.1.0
python-dotenv==1.0.0
django-crispy-forms==2.1
crispy-bootstrap4==2023.1
django-allauth==0.61.0 